Django/Spotify(Almost Whole by Myself)/Project/requirements.txt
django-4.0
django-simple-search-1.0.2
Pillow-8.4.0